requirements.txt
requests
Flask==3.1.3
supabase==2.28.3
python-dotenv==1.2.2
